Python script to parse PCAP files
This simple script parses PCAP files using the Python library pypcapfile (supplemented with files from the pypcapfile directory) and prints information about the first 15 packets in each PCAP file.  The PCAP filenames are hardcoded into the script, but the comments in the script tell you where to find these files.